+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Components can have linked components in the editor.
 * Added a *clear* button for referential value editors that can be null.
 * MacabreGame object now fires an event when the viewport size changes.
+* Event now fires off when a component is initialized.
 
 ### Refactors